requirements.txt
flake8
django 
django-rest-framework
psycopg2-binary
